refactor(github): move GithubDeployment to  migration's archived package
