+++ backend/plugins/github/models/migrationscripts/20230913_add_tool_demployments.go
@@ -20,40 +20,14 @@ package migrationscripts
 import (
 	"github.com/apache/incubator-devlake/core/context"
 	"github.com/apache/incubator-devlake/core/errors"
-	"github.com/apache/incubator-devlake/core/models/migrationscripts/archived"
-	"time"
+	"github.com/apache/incubator-devlake/plugins/github/models/migrationscripts/archived"
 )
-
-type GithubDeployment struct {
-	archived.NoPKModel `json:"-" mapstructure:"-"`
-	ConnectionId       uint64    `json:"connection_id" gorm:"primaryKey"`
-	GithubId           int       `json:"github_id" gorm:"type:varchar(255)"`
-	Id                 string    `json:"id" gorm:"type:varchar(255);primaryKey"`
-	DatabaseId         uint      `json:"database_id"`
-	CommitOid          string    `json:"commit_oid" gorm:"type:varchar(255)"`
-	Description        string    `json:"description" gorm:"type:varchar(255)"`
-	Environment        string    `json:"environment" gorm:"type:varchar(255)"`
-	State              string    `json:"state" gorm:"type:varchar(255)"`
-	LatestStatusState  string    `json:"latest_status_state" gorm:"type:varchar(255)"`
-	LatestUpdatedDate  time.Time `json:"latest_status_update_date"`
-	RepositoryID       string    `json:"repository_id" gorm:"type:varchar(255)"`
-	RepositoryName     string    `json:"repository_name" gorm:"type:varchar(255)"`
-	RepositoryUrl      string    `json:"repository_url" gorm:"type:varchar(255)"`
-	RefName            string    `json:"ref_name" gorm:"type:varchar(255)"`
-	Payload            string    `json:"payload" gorm:"type:text"`
-	CreatedDate        time.Time `json:"created_at"`
-	UpdatedDate        time.Time `json:"updated_at"`
-}
-
-func (GithubDeployment) TableName() string {
-	return "_tool_github_deployments"
-}
 
 type addDeploymentTable struct {
 }
 
 func (*addDeploymentTable) Up(basicRes context.BasicRes) errors.Error {
-	return basicRes.GetDal().AutoMigrate(&GithubDeployment{})
+	return basicRes.GetDal().AutoMigrate(&archived.GithubDeployment{})
 }
 
 func (*addDeploymentTable) Version() uint64 {
